Refactor flightsql server
