+++ src/main.rs
@@ -1,10 +1,9 @@
-use arrow_flight::flight_service_server::FlightServiceServer;
 use clap::{Args, Command};
 use deltaquery::compute::register_compute_factory;
 use deltaquery::computes::duckdb::DQDuckDBComputeFactory;
 use deltaquery::configs::DQConfig;
-use deltaquery::servers::simple::FlightSqlServiceSimple;
-use deltaquery::servers::single::FlightSqlServiceSingle;
+use deltaquery::servers::flightsql;
+use deltaquery::servers::flightsql::{FlightSqlServiceSimple, FlightSqlServiceSingle};
 use deltaquery::state::DQState;
 use deltaquery::storage::register_storage_factory;
 use deltaquery::storages::delta::DQDeltaStorageFactory;
@@ -16,8 +15,6 @@ use std::time::Duration;
 use std::time::Instant;
 use tokio::sync::Mutex;
 use tokio::time;
-use tonic::transport::Server;
-use tonic::transport::{Certificate, Identity, ServerTlsConfig};
 
 #[derive(Debug, Args)]
 pub struct DQOption {
@@ -95,34 +92,22 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let state = Arc::new(Mutex::new(DQState::new(config.clone()).await));
     handle_state(state.clone());
 
-    let mut server = if let Some(tls_config) = config.tls.as_ref() {
-        let server_cert = std::fs::read_to_string(tls_config.server_cert.clone())?;
-        let server_key = std::fs::read_to_string(tls_config.server_key.clone())?;
-        let client_cert = std::fs::read_to_string(tls_config.client_cert.clone())?;
-
-        let tls_config = ServerTlsConfig::new()
-            .identity(Identity::from_pem(&server_cert, &server_key))
-            .client_ca_root(Certificate::from_pem(&client_cert));
-
-        Server::builder().tls_config(tls_config)?
-    } else {
-        Server::builder()
+    match config.server.as_str() {
+        "single" => {
+            flightsql::server::serve(
+                config,
+                FlightSqlServiceSingle::new(state.clone(), catalog).await,
+            )
+            .await?;
+        }
+        _ => {
+            flightsql::server::serve(
+                config,
+                FlightSqlServiceSimple::new(state.clone(), catalog).await,
+            )
+            .await?;
+        }
     };
-
-    let router = match config.server.as_str() {
-        "single" => server.add_service(FlightServiceServer::new(
-            FlightSqlServiceSingle::new(state.clone(), catalog).await,
-        )),
-        _ => server.add_service(FlightServiceServer::new(
-            FlightSqlServiceSimple::new(state.clone(), catalog).await,
-        )),
-    };
-
-    let listen = config.listen.parse()?;
-
-    log::info!("listening on {:?}", listen);
-
-    router.serve(listen).await?;
 
     Ok(())
 }
