+++ src/servers/flightsql/mod.rs
@@ -1,0 +1,7 @@
+pub mod server;
+pub use simple::FlightSqlServiceSimple;
+pub use single::FlightSqlServiceSingle;
+
+mod helpers;
+mod simple;
+mod single;
