+++ src/servers/mod.rs
@@ -1,23 +1,1 @@
-use arrow_flight::sql::{Any, ProstMessageExt};
-
-pub mod simple;
-pub mod single;
-
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct FetchResults {
-    #[prost(string, tag = "1")]
-    pub handle: ::prost::alloc::string::String,
-}
-
-impl ProstMessageExt for FetchResults {
-    fn type_url() -> &'static str {
-        "type.googleapis.com/arrow.flight.protocol.sql.FetchResults"
-    }
-
-    fn as_any(&self) -> Any {
-        Any {
-            type_url: FetchResults::type_url().to_string(),
-            value: ::prost::Message::encode_to_vec(self).into(),
-        }
-    }
-}
+pub mod flightsql;
